requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.5
dj-rest-auth==2.2.3
Django==4.0.2
django-allauth==0.48.0
django-cache-url==3.3.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-model-utils==4.2.0
djangorestframework==3.13.1
djfernet==0.8.1
environs==9.5.0
idna==3.3
marshmallow==3.14.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
